Change tribool to enum class and move to separate header
Also add or_tribool and orwk_tribool for completeness
