+++ symengine/tests/basic/test_basic.cpp
@@ -47,7 +47,6 @@ using SymEngine::sdiff;
 using SymEngine::set_basic;
 using SymEngine::Symbol;
 using SymEngine::symbol;
-using SymEngine::tribool;
 using SymEngine::umap_basic_basic;
 using SymEngine::umap_basic_num;
 using SymEngine::unified_compare;
@@ -1127,51 +1126,4 @@ TEST_CASE("args: Basic", "[basic]")
 
     r1 = log(pi);
     REQUIRE(vec_basic_eq_perm(r1->get_args(), {pi}));
-}
-
-TEST_CASE("tribool", "[basic]")
-{
-    REQUIRE(!is_true(tribool::indeterminate));
-    REQUIRE(!is_true(tribool::trifalse));
-    REQUIRE(is_true(tribool::tritrue));
-
-    REQUIRE(!is_false(tribool::indeterminate));
-    REQUIRE(is_false(tribool::trifalse));
-    REQUIRE(!is_false(tribool::tritrue));
-
-    REQUIRE(is_indeterminate(tribool::indeterminate));
-    REQUIRE(!is_indeterminate(tribool::trifalse));
-    REQUIRE(!is_indeterminate(tribool::tritrue));
-
-    REQUIRE(is_false(and_tribool(tribool::trifalse, tribool::trifalse)));
-    REQUIRE(is_false(and_tribool(tribool::trifalse, tribool::indeterminate)));
-    REQUIRE(is_false(and_tribool(tribool::trifalse, tribool::tritrue)));
-    REQUIRE(is_false(and_tribool(tribool::indeterminate, tribool::trifalse)));
-    REQUIRE(is_false(and_tribool(tribool::tritrue, tribool::trifalse)));
-    REQUIRE(is_indeterminate(
-        and_tribool(tribool::indeterminate, tribool::indeterminate)));
-    REQUIRE(is_indeterminate(
-        and_tribool(tribool::tritrue, tribool::indeterminate)));
-    REQUIRE(is_indeterminate(
-        and_tribool(tribool::indeterminate, tribool::tritrue)));
-    REQUIRE(is_true(and_tribool(tribool::tritrue, tribool::tritrue)));
-
-    REQUIRE(is_true(not_tribool(tribool::trifalse)));
-    REQUIRE(is_false(not_tribool(tribool::tritrue)));
-    REQUIRE(is_indeterminate(not_tribool(tribool::indeterminate)));
-
-    REQUIRE(is_false(andwk_tribool(tribool::trifalse, tribool::trifalse)));
-    REQUIRE(is_indeterminate(
-        andwk_tribool(tribool::trifalse, tribool::indeterminate)));
-    REQUIRE(is_false(andwk_tribool(tribool::trifalse, tribool::tritrue)));
-    REQUIRE(is_indeterminate(
-        andwk_tribool(tribool::indeterminate, tribool::trifalse)));
-    REQUIRE(is_false(andwk_tribool(tribool::tritrue, tribool::trifalse)));
-    REQUIRE(is_indeterminate(
-        andwk_tribool(tribool::indeterminate, tribool::indeterminate)));
-    REQUIRE(is_indeterminate(
-        andwk_tribool(tribool::tritrue, tribool::indeterminate)));
-    REQUIRE(is_indeterminate(
-        andwk_tribool(tribool::indeterminate, tribool::tritrue)));
-    REQUIRE(is_true(andwk_tribool(tribool::tritrue, tribool::tritrue)));
 }
